fix: anchor position now correctly tracked
Anchor position wasn't being correctly tracked because Cb() pulls from
the *outermost named capture*. so all of the named captures in nested
captures, such as lists and strings, were not updated on newlines and so
anchors were unreliable. the grammar can now handle this properly and
transmit context. it should be significantly easier to transmit
additional context (if needed) in the future.

there's now a more general property test for validating outputs of the
format, and some new tests for specific edge cases.

comments and longstrings now normalize their indentation levels to the
indentation level of the parent.