wip
current issue: the tuple type comparer can't see what the element values
is, so it can't compare correctly. open's idea to fix this is to add a
most-specific-type type family, that takes a type T and a value x of
type T, and is a subtype of T. tuple constructions should then produce
types that incorporate this type family, and tuple compare gets to use
the values of the elements when checking.