Attempt to fix neutral value issues (doesn't work)
so when comparing tuple descs (possibly via comparing tuple types) we
don't get the tuple value to help evaluate the element types. so a
free.unique value was used, but this caused conflict when the tuple desc
is dependent (i.e. an element type uses part of the tuple value) and a
later constraint eventually tried to constrain it with the actual tuple
value (usually a lambda argument placeholder).

this commit attempts to use substitution to resolve this issue. another
possibility is to use metavariables.