[WIP] add `host-array-from-tuple` and `make-host-array`
For easier host array construction.
`host-array-from-tuple` directly takes a tuple (uncurried);
`make-host-array` returns a function that takes a tuple (curried).

Ideally, one would be implemented in terms of the other,
but that's currently broken due to unknown reasons.

I know something is broken,
and it effects use,
but I don't know what.